requirements.txt
# --- For-Bot-Working --------- #
pyrogram
TgCrypto
pyromod==1.5
python-dotenv
requests  # Added for making HTTP requests
# --- For-Database ------------ #
pymongo
dnspython
# --- For-Web-Response ------- #
aiohttp